add snowflake ids, instance timeouts, and database API server

+++ Assets/AnyRPG/Core/System/Scripts/DataModels/ItemInstanceSaveData.cs
@@ -10,7 +10,7 @@ namespace AnyRPG {
         public string DisplayName = string.Empty;
         public string ItemQuality = string.Empty;
         public int DropLevel;
-        public int ItemInstanceId;
+        public long ItemInstanceId;
 
         public List<int> RandomSecondaryStatIndexes = new List<int>();
         public string GainCurrencyName;
