+++ Assets/AnyRPG/Core/System/Scripts/Items/Bag.cs
@@ -22,7 +22,7 @@ namespace AnyRPG {
             }
         }
 
-        public override InstantiatedItem GetNewInstantiatedItem(SystemGameManager systemGameManager, int itemId, Item item, ItemQuality usedItemQuality) {
+        public override InstantiatedItem GetNewInstantiatedItem(SystemGameManager systemGameManager, long itemId, Item item, ItemQuality usedItemQuality) {
             if ((item is Bag) == false) {
                 return null;
             }
