+++ Assets/AnyRPG/Core/System/Scripts/Items/CurrencyItem.cs
@@ -23,7 +23,7 @@ namespace AnyRPG {
         public string GainCurrencyName { get => gainCurrencyName; }
         public int GainCurrencyAmount { get => gainCurrencyAmount; }
 
-        public override InstantiatedItem GetNewInstantiatedItem(SystemGameManager systemGameManager, int itemId, Item item, ItemQuality usedItemQuality) {
+        public override InstantiatedItem GetNewInstantiatedItem(SystemGameManager systemGameManager, long itemId, Item item, ItemQuality usedItemQuality) {
             if ((item is CurrencyItem) == false) {
                 return null;
             }
